requirements.txt
pandas
numpy
google-cloud-storage
scikit-
pyyaml